separate puppeteer/dappeteer setup code
