+++ e2e/constants.js
@@ -1,0 +1,8 @@
+'use strict'
+module.exports = {
+  // hardhat default seed
+  // https://hardhat.org/hardhat-network/#hardhat-network-initial-state
+  seed: 'test test test test test test test test test test test junk',
+  // metamask address from seed
+  address: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266'
+}
